bug #8592 - allow leading zeros in p/o numbers; needed to change enterReceipt so it treats order numbers as text fields and use an OrderCluster instead of a PoLineEdit and explicit ellipsis button on the voucher window.

+++ xtupleserver/trunk/dbscripts/functions/enterreceipt.sql
@@ -109,7 +109,7 @@ BEGIN
       recv_notes, recv_freight, recv_freight_curr_id
     ) VALUES (
       _recvid, _timestamp,
-      _o.orderhead_number::INTEGER, pordertype, _o.orderitem_id::INTEGER,
+      _o.orderhead_number, pordertype, _o.orderitem_id::INTEGER,
       CURRENT_USER, _o.orderhead_agent_username, _o.itemsite_id::INTEGER,
       _o.vend_id::INTEGER, _o.vend_item_number, _o.vend_item_descrip,
       _o.vend_uom, pQty, _o.duedate,
